Agregacion de estilo dark

+++ examplevaadin/src/main/java/pe/com/vical/examplevaadin/dao/IParametroDao.java
@@ -2,6 +2,6 @@ package pe.com.vical.examplevaadin.dao;
 
 import pe.com.vical.examplevaadin.domain.Parametro;
 
-public interface IParametroDao extends IBaseDAO<Parametro, Long> {
+public interface IParametroDao extends IBaseDAO<Parametro, Integer> {
 
 }
