+++ examplevaadin/src/main/java/pe/com/vical/examplevaadin/dao/IValorDao.java
@@ -2,6 +2,6 @@ package pe.com.vical.examplevaadin.dao;
 
 import pe.com.vical.examplevaadin.domain.Valor;
 
-public interface IValorDao extends IBaseDAO<Valor, Long> {
+public interface IValorDao extends IBaseDAO<Valor, Integer> {
 
 }
