+++ examplevaadin/src/main/java/pe/com/vical/examplevaadin/dao/hibernate/ParametroDao.java
@@ -6,7 +6,7 @@ import pe.com.vical.examplevaadin.dao.IParametroDao;
 import pe.com.vical.examplevaadin.domain.Parametro;
 
 @Repository
-public class ParametroDao extends BaseDAOImpl<Parametro, Long> implements IParametroDao {
+public class ParametroDao extends BaseDAOImpl<Parametro, Integer> implements IParametroDao {
 
 	private static final long serialVersionUID = -6997610036225453638L;
